optmiizations

+++ nars_logic/src/main/java/nars/nal/tlink/TermLinkBuilder.java
@@ -215,7 +215,7 @@ public class TermLinkBuilder extends BagActivator<TermLinkKey,TermLink> implemen
     }
 
     public static final int hash(byte[] prefix, Term target, short type) {
-        return Util.hash(Arrays.hashCode(prefix), target) ^ (37 * type);
+        return Util.hash(Arrays.hashCode(prefix), target) ^ (37 * ((int)type));
     }
